backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
supabase>=2.3.0
aptos-sdk>=0.9.0
pillow>=10.0.0
qrcode
google-auth>=2.22.0